refactor: MenuEntity, AddressEntity를 Menu, Address로 rename 및 패키지 구조 정리
